first round of porting
svn path=/trunk/KDE/kdegames/libkdegames/; revision=466923

+++ highscore/kexthighscore_tab.cpp
@@ -235,7 +235,7 @@ HistogramTab::HistogramTab(QWidget *parent)
     const Item *sitem = internal->scoreInfos().item("score")->item();
     const PlayerInfos &pi = internal->playerInfos();
     const QVector<uint> &sh = pi.histogram();
-    for (uint k=1; k<pi.histoSize(); k++) {
+    for (int k=1; k<( int )pi.histoSize(); k++) {
         QString s1 = sitem->pretty(0, sh[k-1]);
         QString s2;
         if ( k==sh.size() ) s2 = "...";
